Add controller and routes for accounts, restructure account files
